Dividing settings tests
